copilot.ai/requirements.txt
click==8.1.7
fsspec==2024.6.1
schema==0.7.5
litellm>=1.72.0
langsmith>=0.3.18,<0.4.0
httpx==0.28.1
fastapi
uvicorn
python-dotenv
crewai
crewai-tools
pydantic-ai
copilotkit[crewai]
copilotkit
google-generativeai
tensorflow
numpy
packaging
pdfminer.six
polyfile-weave
tf-keras
tokenizers
transformers
poetry
